requirements.txt
faker
boto3
pyyaml